requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
autobahn==19.11.2
Automat==25.4.16
beautifulsoup4==4.13.4
bidict==0.23.1
breeze_connect==1.0.63
bs4==0.0.2
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
constantly==23.10.4
cryptography==45.0.5
distro==1.9.0
dotenv==0.9.9
fastapi==0.116.0
ffmpy==0.6.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
google-auth==2.40.3
google-genai==1.24.0
gradio==5.35.0
gradio_client==1.10.4
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
incremental==24.7.2
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiteconnect==5.0.1
litellm==1.74.0.post1
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.3
numpy==2.3.1
openai==1.93.1
orjson==3.10.18
packaging==25.0
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
ruff==0.12.2
safehttpx==0.1.6
semantic-version==2.10.0
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smolagents==1.19.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tenacity==8.5.0
tiktoken==0.9.0
tokenizers==0.21.2
tomlkit==0.13.3
tqdm==4.67.1
Twisted==25.5.0
txaio==25.6.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
zope.interface==7.2
